Switch MapWithAI to PMTiles
